+++ Annex/Changes.md
@@ -10,6 +10,8 @@
 - **ApiClient.PreparedRequest**
   - Added multiple new functions for parsing optional values 
     (i.e. parse variants that replace empty responses with None but parse non-empty responses)
+- **Future** (**RequestResultExtensions**)
+  - Added `.tryFlatMapSuccess(...)` for asynchronous **RequestResults**
 - **ResponseParser** (**ResponseParseExtensions**)
   - Added `.rightToResponse(...)` for **ResponseParsers** of type **Either**
 - **RequestQueue**